Mental_Health_Chatbot/requirements.txt
langchain
llama-index
haystack

fastapi
uvicorn

requests
python-dotenv

autogpt
babyagi